created a new component for the draft list
