+++ client/src/pages/Drafts/Drafts.test.js
@@ -4,6 +4,8 @@ import { rest } from "msw";
 
 import { resourceStub, server } from "../../../setupTests";
 
+import DraftList from "./DraftList";
+
 import Drafts from "./index";
 
 describe("Drafts", () => {
@@ -31,6 +33,46 @@ describe("Drafts", () => {
 		await expect(
 			screen.findByRole("heading", { level: 3 })
 		).resolves.toHaveTextContent("foo");
+	});
+
+	it("shows the resource", () => {
+		const publish = jest.fn();
+		const resource = resourceStub({
+			description:
+				'Comprehensive guide to setting up the various types of inputs with React (a.k.a. "data binding")',
+			title: "Data Binding in React",
+			url: "https://www.joshwcomeau.com/react/data-binding/",
+		});
+
+		render(<DraftList publish={publish} resources={[resource]} />);
+		expect(screen.getByRole("heading", { level: 3 })).toHaveTextContent(
+			resource.title
+		);
+		expect(
+			screen.getByRole("link", { name: "joshwcomeau.com" })
+		).toHaveAttribute("href", resource.url);
+		expect(screen.getByText(/comprehensive guide/i)).toBeInTheDocument();
+	});
+
+	it("shows a publish button if enabled", async () => {
+		const publish = jest.fn();
+		const resource = resourceStub();
+		const user = userEvent.setup();
+		server.use(
+			rest.get("/api/topics", (req, res, ctx) => {
+				const mockTopics = [
+					{ id: "1", name: "Topic 1" },
+					{ id: "2", name: "Topic 2" },
+				];
+				return res(ctx.json(mockTopics));
+			})
+		);
+
+		render(<DraftList publish={publish} resources={[resource]} />);
+
+		await user.click(screen.getByRole("button", { name: /publish/i }));
+
+		expect(publish).toHaveBeenCalledWith(resource.id);
 	});
 
 	it("lets those resources be published", async () => {
