+++ client/src/pages/Home/index.js
@@ -8,17 +8,14 @@ export function Home() {
 	const resourceService = useService(ResourceService);
 	const searchParams = useSearchParams();
 	const [{ lastPage, resources } = {}, setEnvelope] = useState();
-	const allowTopicFiltering = true;
+
 	useEffect(() => {
 		resourceService.getPublished(searchParams).then(setEnvelope);
 	}, [resourceService, searchParams]);
 
 	return (
 		<section>
-			<ResourceList
-				resources={resources ?? []}
-				allowTopicFiltering={allowTopicFiltering}
-			/>
+			<ResourceList resources={resources ?? []} />
 			<Pagination lastPage={lastPage ?? 1} />
 		</section>
 	);
